get_coordinates not incorporates raster orientation
To avoid confusion, we will always re-orient the image when asking for
coordinates through get_coordinates. The orient_coordinates method is
now private since this is not needed as a part of the interface.

If one wants a different raster orientation, this canbe set throught
the orientation argument to get_coordinates.

Other (smaller) changes:
- absolute coordinates are now always positive.
- added get_tth method to get a twotheta x-axis (might be useful in
  some edge-cases)